dia05/requirements.txt
django==3.2
djangorestframework
django-cors-headers
djangorestframework-simplejwt